callsubmissionpython/requirements.txt
Flask
flask-cors
azure-storage-blob
